task8. JFR sampling
