--- moryakovdv/task8/api/pom.xml
@@ -1,24 +1,0 @@
-<!-- Дочерний POM от task4-->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>otus.moryakovdv.task8.api</groupId>
-	<artifactId>api</artifactId>
-
-	<parent>
-		<groupId>otus.moryakovdv</groupId>
-		<artifactId>task8</artifactId>
-		<version>0.0.1-SNAPSHOT</version>
-	</parent>
-	<dependencies>
-		
-
-
-	</dependencies>
-
-	<build>
-		<finalName>api</finalName>
-	</build>
-</project>
