--- moryakovdv/task8/api/src/main/java/otus/moryakovdv/task8/model/User.java
@@ -1,51 +1,0 @@
-package otus.moryakovdv.task7.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.persistence.Transient;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**Сущность Пользователь*/
-@Entity
-@Table(name = "USERS")
-@NoArgsConstructor
-@EqualsAndHashCode(of = {"id"})
-@Getter
-@Setter
-public class User {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
-	
-	@Column(name = "username")
-	private String userName;
-	
-	@Column(name = "password")
-	private String password;
-	
-	@Transient
-	/**Сессия юзера*/
-	private String sessionId;
-
-	/**Конструктор по логину и паролю*/
-	public User(String userName, String password) {
-		this.userName=userName;
-		this.password=password;
-	}
-
-
-
-	
-	
-
-	
-
-}
