--- moryakovdv/task8/api/src/main/java/otus/moryakovdv/task8/repository/UsersCrudRepository.java
@@ -1,19 +1,0 @@
-package otus.moryakovdv.task7.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import otus.moryakovdv.task7.model.User;
-
-/**Spring JPA репозиторий для работы с сущностью USER*/
-@Repository
-public interface UsersCrudRepository extends CrudRepository<User, Long> {
-	
-	Optional<User> findByUserNameAndPassword(String userName, String password);
-
-	
-	
-	
-}
