--- moryakovdv/task8/api/src/main/java/otus/moryakovdv/task8/service/LoginFailedException.java
@@ -1,18 +1,0 @@
-package otus.moryakovdv.task7.service;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-/**Кастомное исключение - не авторизован*/
-@ResponseStatus(value = HttpStatus.UNAUTHORIZED, reason = "Login failed")
-public class LoginFailedException extends RuntimeException {
-
-	
-	
-	private static final long serialVersionUID = 1937936107282823445L;
-
-
-
-  
-
-}
