--- moryakovdv/task8/api/src/main/java/otus/moryakovdv/task8/service/PasswordHasher.java
@@ -1,11 +1,0 @@
-package otus.moryakovdv.task7.service;
-
-import java.io.UnsupportedEncodingException;
-import java.security.NoSuchAlgorithmException;
-
-/**Общий интерфйс для классов-хешировщиков*/
-public interface PasswordHasher {
-
-	String hash(String input) throws UnsupportedEncodingException, NoSuchAlgorithmException;
-	
-}
