--- moryakovdv/task8/api/src/test/resources/application.properties
@@ -1,7 +1,0 @@
-#spring.datasource.url=jdbc:h2:mem:testdb
-#spring.jpa.hibernate.ddl-auto=create-drop
-
-spring.datasource.url=jdbc:h2:file:./db-task7;DB_CLOSE_ON_EXIT=TRUE
-spring.jpa.hibernate.ddl-auto=create-drop
-
-app.help.message = <h2>USAGE:</h2><p>/user/login - do a login with provided userName and passwd parameters
