+++ moryakovdv/task8/api8/src/main/resources/LICENSE
