+++ moryakovdv/task8/api8/src/test/java/otus/moryakovdv/task8/ApplicationMockingTests.java
@@ -1,0 +1,69 @@
+/*
+ * Copyright 2012-2014 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package otus.moryakovdv.task8;
+
+import static org.hamcrest.Matchers.containsString;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.ApplicationContext;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+
+import otus.moryakovdv.task8.service.PasswordHasher;
+
+
+@SpringBootTest()
+@AutoConfigureMockMvc
+class ApplicationMockingTests {
+
+	@Autowired
+	private MockMvc mockMvc;
+	
+	@Autowired
+	private ApplicationContext context;
+	
+	private PasswordHasher hasher;
+
+	@Test
+	void testDefaultMessage() throws Exception {
+		this.mockMvc.perform(get("/"))
+			.andDo(print())
+			.andExpect(status().isOk())
+			.andExpect(content().string(containsString("USAGE")));
+
+	}
+	
+	@Test
+	public void testGetAllUsers()  throws Exception {
+		
+		MvcResult res = this.mockMvc.perform(get("/userCount"))
+			.andDo(print())
+			.andExpect(status().isOk())
+			.andReturn();
+		
+		
+		
+	}
+	
+}
