+++ moryakovdv/task8/api8/src/test/resources/LICENSE
