--- moryakovdv/task8/embeddedjmeter/src/main/java/module-info.java
@@ -1,6 +1,0 @@
-module otus.moryakovdv.task7.embeddedjmeter {
-	requires ApacheJMeter.core;
-	requires jorphan;
-	requires lombok;
-	requires org.slf4j;
-}
