+++ moryakovdv/task8/embeddedjmeter8/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module otus.moryakovdv.task8.embeddedjmeter {
+	requires ApacheJMeter.core;
+	requires jorphan;
+	requires lombok;
+	requires org.slf4j;
+}
